feat(input): 添加手势控制模式和高级手势设置
1. 在MainWindow.xaml中新增控制模式选择和手势设置面板
2. 重构MainViewModel，支持相对位置和手势模式切换
3. 更新TouchpadPoint模型，优化数据类型和属性
4. 在WindowsInputSimulator中添加虚拟键码枚举和按键模拟方法
5. 实现手势识别、灵敏度调整和手势动作映射功能
6. 支持多种手势动作，如翻页、媒体控制、复制粘贴等
7. 增加手势提示开关和可视化支持

+++ GearVRController/GearVRController/Views/TouchpadVisualizerWindow.xaml.cs
@@ -306,7 +306,7 @@ namespace GearVRController.Views
                 }
                 
                 // 记录历史
-                var point = new TouchpadPoint(processedX, processedY, isPressed);
+                var point = new TouchpadPoint((float)processedX, (float)processedY, isPressed);
                 _touchpadHistory.Add(point);
                 
                 // 维持最大历史点数
